.
Added feature to dispatch client errors to other faculties, so that engineers can respond to it.

Changed the mechanism of session of management to one where sessions are directly loaded from the database, not memory.

Added session regeneration, for increased security.

Fixed minor bugs with json-rpc.

Added the ability for clients on the frontend to provide methods for the server.