.
Updated to the latest version of html-hc, with new features, and bug fixes.

Fixed minor bug with json-rpc cache, that caused data to persist for too long, if reloads happen before expiry.

Created event-based-extender utility, for components, that need functionalities that are provided by other components (the providers). In this way, components can be extended easily, in a very flexible way.

Fixed minor with with the worker-world utility.

Improved typing.